Extract mapping module
